requirements.txt
opencv-python
python-dotenv
numpy
python-telegram-bot
psycopg2
